legger til basepath for å få riktig hosting på nav.no/arbeid
