+++ next.config.js
@@ -1,18 +1,15 @@
 const withLess = require("@zeit/next-less");
 
-const navFrontendModuler = Object.keys(
-  require("./package.json").dependencies
-).filter((it) => it.startsWith("nav-"));
+const navFrontendModuler = Object.keys(require("./package.json").dependencies).filter((it) => it.startsWith("nav-"));
 
-const withTranspileModules = require("next-transpile-modules")(
-  navFrontendModuler
-);
+const withTranspileModules = require("next-transpile-modules")(navFrontendModuler);
 
 module.exports = withTranspileModules(
   withLess({
+    basePath: "/arbeid",
     i18n: {
-      locales: ['no', 'en'],
-      defaultLocale: 'no'
+      locales: ["no", "en"],
+      defaultLocale: "no",
     },
   })
 );
