+++ src/pages/404.tsx
@@ -1,16 +1,16 @@
 import React from "react";
-import SEO from "../../components/SEO";
+import SEO from "../components/SEO";
 import styled from "styled-components/macro";
 import { Innholdstittel, Normaltekst } from "nav-frontend-typografi";
 import { useTranslation } from "react-i18next";
-import SideListe from "../../templates/faktaside/Navigasjonsmeny/SideListe";
+import SideListe from "../templates/faktaside/Navigasjonsmeny/SideListe";
 import { useLocation, useMount } from "react-use";
-import { loggNotFound } from "../../utils/logging";
+import { loggNotFound } from "../utils/logging";
 import { GetStaticProps } from "next";
-import { SupportedLanguage } from "../../i18n/supportedLanguages";
-import fetchFaktasiderMenuData from "../../hooks/graphQl/fetchFaktasiderMenuData";
+import { SupportedLanguage } from "../i18n/supportedLanguages";
+import fetchFaktasiderMenuData from "../hooks/graphQl/fetchFaktasiderMenuData";
 import { useRouter } from "next/router";
-import { MenuItem } from "../../hooks/graphQl/menuDataUtils";
+import { MenuItem } from "../hooks/graphQl/menuDataUtils";
 
 const Style = styled.div`
   display: flex;
@@ -28,8 +28,8 @@ export const getStaticProps: GetStaticProps = async (context) => {
 
   return {
     props: {
-      menuData
-    }
+      menuData,
+    },
   };
 };
