+++ src/pages/demoapp.tsx
@@ -1,26 +1,25 @@
 import * as React from "react";
 import styled from "styled-components/macro";
-import UnderArbeid from "../../templates/veiviser/UnderArbeid";
-import VelgDetSomPasserBest, { VeiviserValg } from "../../templates/veiviser/VelgDetSomPasserBest";
-import { veiviserMachine } from "../../templates/veiviser/VeiviserStateChart";
+import UnderArbeid from "../templates/veiviser/UnderArbeid";
+import VelgDetSomPasserBest, { VeiviserValg } from "../templates/veiviser/VelgDetSomPasserBest";
+import { veiviserMachine } from "../templates/veiviser/VeiviserStateChart";
 import { useMachine } from "@xstate/react";
-import getAlleTilpassInnholdValg from "../../templates/faktaside/TilpassInnhold/getAlleTilpassInnholdValg";
-import { Group, isGroup } from "../../utils/richTextUtils/richTextTypes";
-import BlockContent from "../../components/BlockContent/BlockContent";
-import { visBasertPåFiltrering } from "../../components/BlockContent/VisFor/VisFor";
-import { typografiStyle } from "../../templates/faktaside/MainContentStyle";
-import VeiviserBrødsmuler from "../../templates/veiviser/VeiviserBrødsmuler";
-import DevKnapper from "../../components/DevKnapper/DevKnapper";
-import { useVisForContext } from "../../components/BlockContent/VisFor/VisForContext";
-import { visBasertPaaVisPaaConfig } from "../../components/BlockContent/VisFor/VisPaaSide";
-import { isDevelopment } from "../../utils/environment";
-import { createH2Group } from "../../utils/richTextUtils/createGroup";
+import getAlleTilpassInnholdValg from "../templates/faktaside/TilpassInnhold/getAlleTilpassInnholdValg";
+import { Group, isGroup } from "../utils/richTextUtils/richTextTypes";
+import BlockContent from "../components/BlockContent/BlockContent";
+import { visBasertPåFiltrering } from "../components/BlockContent/VisFor/VisFor";
+import { typografiStyle } from "../templates/faktaside/MainContentStyle";
+import VeiviserBrødsmuler from "../templates/veiviser/VeiviserBrødsmuler";
+import DevKnapper from "../components/DevKnapper/DevKnapper";
+import { useVisForContext } from "../components/BlockContent/VisFor/VisForContext";
+import { visBasertPaaVisPaaConfig } from "../components/BlockContent/VisFor/VisPaaSide";
+import { isDevelopment } from "../utils/environment";
+import { createH2Group } from "../utils/richTextUtils/createGroup";
 import { Knapp } from "nav-frontend-knapper";
 import { GetStaticProps } from "next";
-import { SupportedLanguage } from "../../i18n/supportedLanguages";
-import fetchAllFaktasider from "../../hooks/graphQl/fetchAllFaktasider";
-import { FaktasideContext } from "../../hooks/graphQl/fetchFaktaside";
-
+import { SupportedLanguage } from "../i18n/supportedLanguages";
+import fetchAllFaktasider from "../hooks/graphQl/fetchAllFaktasider";
+import { FaktasideContext } from "../hooks/graphQl/fetchFaktaside";
 
 const Style = styled.div`
   ${typografiStyle};
@@ -47,8 +46,8 @@ export const getStaticProps: GetStaticProps = async (context) => {
 
   return {
     props: {
-      pages: JSON.parse(JSON.stringify(faktasider))
-    }
+      pages: JSON.parse(JSON.stringify(faktasider)),
+    },
   };
 };
 
@@ -62,8 +61,8 @@ function Demoapp(props: Props) {
   const [state, send] = useMachine(veiviserMachine, {
     actions: {
       setFiltrering: (ctx) => visForContest.dispatch({ type: "setKey", key: ctx.filtrering! }),
-      clearFiltrering: () => visForContest.dispatch({ type: "clear" })
-    }
+      clearFiltrering: () => visForContest.dispatch({ type: "clear" }),
+    },
   });
 
   const context = state.context;
@@ -73,37 +72,37 @@ function Demoapp(props: Props) {
   const siderValg: VeiviserValg<FaktasideContext>[] = pages.map((page) => ({
     label: page.title || "Mangler tittel",
     id: page.id,
-    object: page
+    object: page,
   }));
 
   const filtreringsValg: VeiviserValg<string>[] = context.side
     ? getAlleTilpassInnholdValg(context.side.innhold).map((valg) => ({
-      label: valg,
-      id: valg,
-      object: valg
-    }))
+        label: valg,
+        id: valg,
+        object: valg,
+      }))
     : [];
 
   const overskriftsValg: VeiviserValg<Group>[] = context.side
     ? context.side.innhold
-      .filter(isGroup)
-      .filter((group: Group) => visBasertPåFiltrering(visForContest, group.blockConfig?.visFor).vis)
-      .filter((group: Group) =>
-        visBasertPaaVisPaaConfig(state.context.side?.id || "", group.blockConfig?.visPaaSider)
-      )
-      .filter((group: Group) => isDevelopment() || !group.blockConfig?.erUtkast)
-      .map((group) => ({
-        label: group.title,
-        id: group.blockConfig?.id || "N/A",
-        object: group
-      }))
+        .filter(isGroup)
+        .filter((group: Group) => visBasertPåFiltrering(visForContest, group.blockConfig?.visFor).vis)
+        .filter((group: Group) =>
+          visBasertPaaVisPaaConfig(state.context.side?.id || "", group.blockConfig?.visPaaSider)
+        )
+        .filter((group: Group) => isDevelopment() || !group.blockConfig?.erUtkast)
+        .map((group) => ({
+          label: group.title,
+          id: group.blockConfig?.id || "N/A",
+          object: group,
+        }))
     : [];
 
   if (context.side?.kortFortalt) {
     overskriftsValg.unshift({
       label: "Kort fortalt",
       id: "kort-fortalt",
-      object: createH2Group("Kort fortalt", context.side.kortFortalt)
+      object: createH2Group("Kort fortalt", context.side.kortFortalt),
     });
   }
 
@@ -111,7 +110,7 @@ function Demoapp(props: Props) {
     overskriftsValg.push({
       label: "Relatert informasjon",
       id: "relatert-informasjon",
-      object: createH2Group("Relatert informasjon", context.side.relatertInformasjon)
+      object: createH2Group("Relatert informasjon", context.side.relatertInformasjon),
     });
   }
 
@@ -145,7 +144,7 @@ function Demoapp(props: Props) {
             onClick={() =>
               send({
                 type: "VELGOVERSKRIFT",
-                group: overskriftsValg[groupIndex - 1]?.object || overskriftsValg[overskriftsValg.length - 1].object
+                group: overskriftsValg[groupIndex - 1]?.object || overskriftsValg[overskriftsValg.length - 1].object,
               })
             }
           >
@@ -156,7 +155,7 @@ function Demoapp(props: Props) {
             onClick={() =>
               send({
                 type: "VELGOVERSKRIFT",
-                group: overskriftsValg[groupIndex + 1]?.object || overskriftsValg[0].object
+                group: overskriftsValg[groupIndex + 1]?.object || overskriftsValg[0].object,
               })
             }
           >
